+++ Cargo.toml
@@ -62,5 +62,5 @@ esp-backtrace = { version = "0.4.0", features = [
 
 [patch.crates-io]
 # TODO: automate these updates
-esp32c3-hal    = { git = "https://github.com/rustbox/esp-hal", rev = "613f96d" }
-esp-hal-common = { git = "https://github.com/rustbox/esp-hal", rev = "613f96d" }
+esp32c3-hal    = { git = "https://github.com/rustbox/esp-hal", rev = "716934cd" }
+esp-hal-common = { git = "https://github.com/rustbox/esp-hal", rev = "716934cd" }
